Added rejoin functionality.
The bot now rejoins the channel its been kicked from, after a given
period of time. You should set it to around 3–5 seconds, depending on
the server—some servers ban you for joining too many times in a period
of time, so this is a safeguard against pesky OPs that want to get you
K-lined. Of course, if someone wants your bot gone, they'll just ban
it from their channel.